requirements.txt
azure-appconfiguration
python-dotenv
